+++ lib/domain/home/home_page.dart
@@ -13,7 +13,7 @@ import 'package:pilll/util/datetime/day.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../calendar/weekly/weekly_calendar.dart';
+import '../../components/organisms/calendar/weekly_calendar.dart';
 
 GlobalKey<_HomePageState> homeKey = GlobalKey();
